+++ EventManagement/Runtime/Scripts/GameEventHandler.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using FredericRP.GenericSingleton;
+using System;
 using System.Collections.Generic;
 using UnityEngine.Events;