+++ EventManagement/Runtime/Scripts/ValueReference/FloatReference.cs
@@ -1,6 +1,9 @@
 ﻿using UnityEngine;
 
-[CreateAssetMenu(menuName = "FredericRP/Value Reference/Float")]
-public class FloatReference : GenericReference<float>
+namespace FredericRP.EventManagement
 {
+  [CreateAssetMenu(menuName = "FredericRP/Value Reference/Float")]
+  public class FloatReference : GenericReference<float>
+  {
+  }
 }