+++ EventManagement/Runtime/Scripts/ValueReference/IntReference.cs
@@ -1,6 +1,9 @@
 ﻿using UnityEngine;
 
-[CreateAssetMenu(menuName="FredericRP/Value Reference/Integer")]
-public class IntReference : GenericReference<int>
+namespace FredericRP.EventManagement
 {
+  [CreateAssetMenu(menuName = "FredericRP/Value Reference/Integer")]
+  public class IntReference : GenericReference<int>
+  {
+  }
 }