+++ PersistentDataSystem/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 ### Fixed
+- Find and load the correct assembly for the children of SavedData classes to load
+- Editor assembly restricted to Editor platform
 
 ### Added
 - First publish of this new package: a system to load and save persistent data like default player values, player saves, etc.